fetch: add new fetch.default configuration
When the user has an upstream branch configured to track a remote tracking
branch:

  % git checkout --set-upstream-to github/master

Doing a 'git fetch' without any arguments would try to fetch 'github', because
it's configured as current branch's remote (branch.<current>.remote).

However, if we do something slightly different:

  % git checkout --set-upstream-to master

Doing a 'git fetch' without any arguments would try to fetch the remote '.'.

But, unlike the first instance, the second command would not update any remote
tracking branch, in fact, it would not update any branch at all. The only effect
is that it would update FETCH_HEAD, which is not useful in the least.

The vast majority of them would have absolutely no idea what's going on.  Many
of them would probably just shrug, and manually specify the remote the want to
fetch from, which is 'origin' in many cases, *if* they manage to notice the '.'
at all.

What most users would want to do is fetch the default remote as if no upstream
branch was configured, usually "origin".

This would break backwards compatibility (although it's unclear why any user
would want to run `git fetch` and not want to actually fetch anything), so to
introduce the new sane behavior we add the new `fetch.default` configuration.

By default it will be `current`, which maintains the current behavior of
fetching the current remote no matter what (even '.').

But the new behavior of `sane` will ignore the '.' remote, turning `git fetch`
into a useful command once again.

No functional changes.

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>