branch: prioritize upstream in -v
The two important pieces of information relevant to -v and -vv are the
upstream branch, and tracking status (ahead/behind).

Currently we give priority to the tracking stats (shown with -v) and
demote the upstream branch name (shown only with -vv).

However, showing "[ahead 10]" is not useful to most users, since they
will wonder: ahead of what?.

A lot of people set their upstream to the place they push to, not where
they base their branch on. In fact, many guides suggest the following to
set the upstream branch:

  git push --set-upstream github pull-request

Even more, when pushing a fresh branch git itself recommends that:

  % git clone $url .
  % git checkout -b fix
  ...
  % git push

  fatal: The current branch fix has no upstream branch.
  To push the current branch and set the remote as upstream, use

      git push --set-upstream origin fix

Therefore for many workflows `git branch -v` is not very useful.

Inevitably many users will need to know what @{upstream} is.

So let's make `git branch -v` output what is most useful:

  [myrepo/mybranch]

Before:

  * mybranch b2489a3735 [ahead 1] Hot fix

After:

  * mybranch b2489a3735 [myrepo/mybranch] Hot fix

An additional benefit is that `git branch -v` is faster: from 2 to 40
times (depending on the number of stale branches).

`git branch -vv` is unaffected.

Suggestions-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>
Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>