Add concept of 'publish' branch
The publish branch is the branch the user wants to push to, akin to the
upstream branch, which is the branch the user wants to fetch from.

It is the branch the user will broadcast to do pull requests from.

It overrides other configurations, such as remote.pushDefault,
remote.<name>.push, and push.default. All of which are poor attempts at
proving support for triangular workflows.

All the user needs is a branch to fetch from, and a branch to push to.
It's that simple.

The upstream branch is:

  branch.$name.remote
  branch.$name.merge

The publish branch is:

  branch.$name.pushremote
  branch.$name.push

Since there's already a pushremote, all that is missing is the branch
name.

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>