rebase: add REBASE_DEFAULT
By introducing a default we can distinguish when the user has forced an
option.

Therefore there's no need pass around an extra variable (it's the same
as opt_rebase == REBASE_DEFAULT), nor is there any need to initialize
opt_rebase to an invalid value.

Additionally, this will allow us to override the default with a
configuration, and subsequently that configuration with arguments.

Cc: Junio C Hamano <gitster@pobox.com>
Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>