pull: improve default warning
We don't want to start by recommending a permanent configuration, but a
temporary solution so they start training their fingers and maybe learn
how to do a rebase. So we start with the commands.

Also, we need to be clear about what we mean by "specifying"; merge, or
rebase.

Moreover, it's better use --global in the configuration commands like we
did with push.default.

And finally, point to the documentation that explains what is a
non-fast-forward, and how to solve it:

  git help fast-forward

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>