pull: don't consider --ff options especially
We want the user to specify either --merge or --rebase, if she doesn't
we throw a warning.

Using --ff, --no-ff, or --ff-only does not make the merge explicit.

For example, if the user has the following configuration:

  git config pull.rebase true
  git pull --no-ff

A merge is not implied.

We should be consistent and either imply a merge--in which case a
previous "pull.rebase=true" configuration is overridden--or don't--in
which case the warning should be thrown.

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>