test: completion: add zsh tests
zsh's completion system is very advanced, but that comes with a
disadvantage: there is no easy way to dump the completions. The simplest
way of doing that is printing them to stdout in a certain format, and
then use the zpty module to parse that format and actually print them.

There is no easier way.

We also have to simulate `compadd` in order to output the list of
completions in a way that is similar to the bash completion tests.

The resulting test scripts is very close to the bash completion test
script, except some tests don't make sense, and others require minor
modifications.

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>