test: fix build for zsh
zsh has the ability to emulate Bourne shell, which is closer to what our
testing suite expects.

Using the POSIX_ARGZERO option makes $0 behave as POSIX seems to define:
show the name of the command launched, as opposed to the zsh default:
the path of the script.

This enables many tests, but not all.

We can run the tests with `zsh --emulate sh` which has more
compatibility, but doing `emulate sh` inside the script allows us to
just use `zsh`, although it's not as compatible.

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>