config: use strbuf when renaming a section
When renaming a configuration section which has an entry whose length
exceeds the size of our buffer in
`git_config_copy_or_rename_section_in_file()`, git will incorrectly form
a new configuration section with part of the data in the section being
removed.

For example:

    [b]
      c = d <spaces> [a] e = f
    [a]
      g = h

Here, we have two sections with two configuration values. The text
"[a] e = f" belongs to the configuration named "b.c".

However, when renaming the section 'a', git will wrongly use the text
from the "b.c" configuration.

The reason why this happens is that the code is relying on multiple
calls to `fgets()`, and when there isn't enough room in the buffer, an
extra call to `fgets()` will be issued. If the next effective character
of the buffer is a '[', the code will incorrectly consider it the
beginning of a new section.

Let's avoid this by using strbuf instead.

Reported-by: André Baptista <andre@ethiack.com>
Reported-by: Vítor Pinho <vitor@ethiack.com>
Reported-by: Charly Reux <charly.reux@inria.fr>
Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>